// src/app/modules/globalLink/globalLinks.interface.ts
export interface IGlobalLinks {
  facebook_group: string;
  submit_photo: string;
  instagram: string;
  sponsor: string;
  linked_in: string;
  facebook_page: string;
  submission_link: string;
  pagel_email: string;
  pagel_facebook: string;
  pagel_instagram: string;
  pagel_web: string;
  sabbir_email: string;
  sabbir_facebook: string;
  sabbir_instagram: string;
  sabbir_web: string;
  twitter: string;
  Insta_access_token: string;
  pagel_photo: string;
}